Move helper functions into content-helpers.ts
