+++ src/content-helpers.ts
@@ -18,6 +18,9 @@ limitations under the License.
 /** @module ContentHelpers */
 
 import { MsgType } from "./@types/event";
+import { TEXT_NODE_TYPE } from "./@types/extensible_events";
+import { ILocationContent, LOCATION_EVENT_TYPE, TIMESTAMP_NODE_TYPE } from "./@types/location";
+import { IPollContent, POLL_START_EVENT_TYPE } from "./@types/polls";
 
 /**
  * Generates the content for a HTML Message event
@@ -97,5 +100,62 @@ export function makeEmoteMessage(body: string) {
     return {
         msgtype: MsgType.Emote,
         body: body,
+    };
+}
+
+/**
+ * Generates the content for a Poll Start event
+ * @param question the poll question
+ * @param answers the possible answers
+ * @param kind whether the poll is disclosed or undisclosed. Allowed values are
+ *             "m.poll.disclosed" or "m.poll.undisclosed", or the unstable equivalents
+ */
+export function makePollContent(
+    question: string,
+    answers: string[],
+    kind: string,
+): IPollContent {
+    question = question.trim();
+    answers = answers.map(a => a.trim()).filter(a => !!a);
+    return {
+        [TEXT_NODE_TYPE.name]:
+            `${question}\n${answers.map((a, i) => `${i + 1}. ${a}`).join('\n')}`,
+        [POLL_START_EVENT_TYPE.name]: {
+            kind: kind,
+            question: {
+                [TEXT_NODE_TYPE.name]: question,
+            },
+            answers: answers.map(
+                (a, i) => ({ id: `${i}-${a}`, [TEXT_NODE_TYPE.name]: a }),
+            ),
+        },
+    };
+}
+
+/**
+ * Generates the content for a Location event
+ * @param text a text for of our location
+ * @param uri a geo:// uri for the location
+ * @param ts the timestamp when the location was correct (milliseconds since
+ *           the UNIX epoch)
+ * @param description the (optional) label for this location on the map
+ */
+export function makeLocationContent(
+    text: string,
+    uri: string,
+    ts: number,
+    description?: string,
+): ILocationContent {
+    return {
+        "body": text,
+        "msgtype": MsgType.Location,
+        "geo_uri": uri,
+        [LOCATION_EVENT_TYPE.name]: {
+            uri,
+            description,
+        },
+        [TIMESTAMP_NODE_TYPE.name]: ts,
+        [TEXT_NODE_TYPE.name]: text,
+        // TODO: MSC1767 fallbacks m.image thumbnail
     };
 }
